simplify docker setup
Docker is fine for development and also in production environments.
However, the setup should be as simple as possible.
It's now possible to use only one command and setup a complete testing/
development environment.
It's not intended to be used in production, but this can be easily used as a
starting point for a complete setup.