rework internal actor API
this API used to look like some real steps. However, this is more like
what is done by ApiDataFactory of codeceptjs. Currently the
implementation still uses the actor to create stuff, but this might
change in the future. At least, now there is the option to directly
manipulate data using the REST API